Update dashboard and language screens with new appbar
