+++ app/src/main/java/org/greenstand/android/TreeTracker/view/Buttons.kt
@@ -1,0 +1,26 @@
+package org.greenstand.android.TreeTracker.view
+
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.Button
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun TextButton(
+    modifier: Modifier = Modifier,
+    stringRes: Int,
+    onClick: () -> Unit,
+) {
+    // TODO customize button visuals
+    Button(
+        onClick = onClick,
+        modifier = modifier.size(height = 46.dp, width = 110.dp)
+    ) {
+        Text(
+            text = stringResource(id = stringRes)
+        )
+    }
+}
